update : 로그인 검증 로직 만듬

+++ src/main/java/pakaCoding/flower/controller/CartController.java
@@ -87,6 +87,8 @@ public class CartController {
 
         if(count < 1){
             return new ResponseEntity<String>("최소 1개 이상 담아주세요", HttpStatus.BAD_REQUEST);
+        } else if(!cartService.validateCartItem(cartItemId, principal.getName())){
+            return new ResponseEntity<String>("수정 권한이 없습니다.", HttpStatus.FORBIDDEN);
         }
         cartService.updateCartItemCount(cartItemId, count);
         return new ResponseEntity<Long>(cartItemId, HttpStatus.OK);
@@ -96,7 +98,11 @@ public class CartController {
     //장바구니 삭제
     @DeleteMapping(value = "/cartItem/{cartItemId}")
     @ResponseBody
-    public ResponseEntity deleteCartItem(@PathVariable(value = "cartItemId") Long cartItemId){
+    public ResponseEntity deleteCartItem(@PathVariable(value = "cartItemId") Long cartItemId,Principal principal){
+        if(!cartService.validateCartItem(cartItemId, principal.getName())){
+            return new ResponseEntity<String>("수정 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
         cartService.deleteCartItem(cartItemId);
         return new ResponseEntity<Long>(cartItemId, HttpStatus.OK);
     }
